Content Edge Code/Content Edge Codes/requirements.txt
numpy==2.2.6
pandas==2.3.3
playwright==1.55.0
python-dateutil==2.9.0.post0
typing_extensions==4.15.0
boto3==1.40.75
python-dotenv==1.1.1
